Update Main Menu UI layout scaling, customize button/logo sizes, and sync project workspace changes

+++ Assets/Scenes/scripts/Tsuchigumo Boss/SpiderBoss.cs
@@ -63,7 +63,7 @@ public class SpiderBoss : MonoBehaviour
 
     private void UpdatePhase()
     {
-        float ratio = health.CurrentHealth / health.MaxHealth;
+        float ratio = (float)health.CurrentHealth / health.MaxHealth;
         int phase = 1;
         if (ratio <= stageThreeHealthThreshold)
             phase = 3;
